Made only the requester (Or me) able to remove bot's messages

+++ src/main/java/com/uddernetworks/emojide/discord/commands/manager/EmbedUtils.java
@@ -10,9 +10,14 @@ import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
 import javax.annotation.Nonnull;
 import java.awt.*;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.function.Consumer;
 
 public class EmbedUtils extends ListenerAdapter {
+
+    private static Map<Message, Long> messageRequesters = Collections.synchronizedMap(new HashMap<>());
 
     public static void error(TextChannel channel, Member author, String message) {
         EmbedBuilder eb = new EmbedBuilder();
@@ -23,7 +28,10 @@ public class EmbedUtils extends ListenerAdapter {
         eb.setDescription(message);
 
         eb.setFooter("In response to " + author.getEffectiveName(), author.getUser().getAvatarUrl());
-        channel.sendMessage(eb.build()).submit().thenAccept(sent -> sent.addReaction("U+1F5D1").queue());
+        channel.sendMessage(eb.build()).submit().thenAccept(sent -> {
+            sent.addReaction("U+1F5D1").queue();
+            messageRequesters.put(sent, author.getIdLong());
+        });
     }
 
     public static Message sendEmbed(TextChannel channel, Member author, String title, String description) {
@@ -33,6 +41,7 @@ public class EmbedUtils extends ListenerAdapter {
     public static Message sendEmbed(TextChannel channel, Member author, String title, Consumer<EmbedBuilder> embedBuilderConsumer) {
         var message = channel.sendMessage(createEmbed(author, title, embedBuilderConsumer)).complete();
         message.addReaction("U+1F5D1").queue();
+        messageRequesters.put(message, author.getIdLong());
         return message;
     }
 
@@ -53,6 +62,12 @@ public class EmbedUtils extends ListenerAdapter {
         if (event.getUser().isBot()) return;
         if (!event.getReactionEmote().getAsCodepoints().equalsIgnoreCase("U+1F5D1")) return;
         event.getChannel().retrieveMessageById(event.getMessageIdLong()).submit().thenAccept(message -> {
+            var userId = event.getUser().getIdLong();
+
+            if (messageRequesters.containsKey(message) && messageRequesters.get(message) != userId) {
+                if (userId != 249962392241307649L) return;
+            }
+
             if (message.getAuthor().getIdLong() == event.getJDA().getSelfUser().getIdLong()) {
                 message.delete().queue();
             }
